Add new SVG assets, update application styles and scripts, and create environment settings
This commit adds new SVG files as well as updates to existing CSS scripts. It also introduces essential project settings via the `.idea` and `eslint` config files. Additionally, React components for the application's main functionality are created along with corresponding styles. Changes applied are part of the initial setup and creation of the web application's structure and design.