Update Hero component with new images and content
Incorporated new imagery and respective content into the Hero component in the React app. Also, the useState function was imported from the React library to replace React.useState and image content was adjusted for a better user experience. Further, the presentation layout was slightly adjusted to house the additional content and provide a more aesthetic feel.