Add Projects section and improve UI components
Created a new 'Projects' section and added it to the Home page. Updated UI components including ActiveCard, InactiveCard, and Numbering, improving the design and adding export structure. The changes also include minor updates to UI colors in Hero and AdsBar components, and the addition of new data and images to a data.js file for utilization in the new section.